Split app.py into four files for better coding practices.
